reorganising models for share app
