+++ backend/tracim_backend/app_models/validator.py
@@ -12,10 +12,10 @@ from marshmallow.validate import Regexp
 from tracim_backend.app_models.contents import GlobalStatus
 from tracim_backend.app_models.contents import content_status_list
 from tracim_backend.app_models.contents import content_type_list
+from tracim_backend.applications.share.models import ContentShare
 from tracim_backend.exceptions import TracimValidationFailed
 from tracim_backend.models.auth import Profile
 from tracim_backend.models.auth import User
-from tracim_backend.models.content_share import ContentShare
 from tracim_backend.models.data import ActionDescription
 from tracim_backend.models.data import UserRoleInWorkspace
 from tracim_backend.views.agenda_api.models import AgendaType
