+++ backend/tracim_backend/tests/functional/test_share_content.py
@@ -2,11 +2,11 @@ import pytest
 from rq import SimpleWorker
 import transaction
 
-from tracim_backend.applications.share.share import ShareLib
+from tracim_backend.applications.share.lib import ShareLib
+from tracim_backend.applications.share.models import ContentShareType
 from tracim_backend.error import ErrorCode
 from tracim_backend.lib.utils.utils import get_redis_connection
 from tracim_backend.lib.utils.utils import get_rq_queue
-from tracim_backend.models.content_share import ContentShareType
 from tracim_backend.models.revision_protection import new_revision
 from tracim_backend.tests.fixtures import *  # noqa: F403,F40
